selecttest/getSelection.html

<!doctype html>
<title>getSelection() tests</title>
<div id=log></div>
<script src=/resources/testharnessreport.js></script>
<script src=http://w3c-test.org/resources/testharness.js></script>
<script>
"use strict";

// TODO: Figure out more places where defaultView is or is not guaranteed to be
// null, and test whether getSelection() is null.

test(function() {
	// Sanity checks like this are to flag known browser bugs with clearer
	// error messages, instead of throwing inscrutable exceptions.
	assert_true("Selection" in window,
		"Sanity check: window must have Selection property");

	assert_true(window.getSelection() instanceof Selection);
}, "window.getSelection() instanceof Selection");

test(function() {
	assert_equals(window.getSelection(), window.getSelection());
}, "window.getSelection() === window.getSelection()");

test(function() {
	assert_true("Selection" in window,
		"Sanity check: window must have Selection property");
	// This sanity check (which occurs a number of times below, too) is because
	// document.getSelection() is supposed to return null if defaultView is
	// null, so we need to figure out whether defaultView is null or not before
	// we can make correct assertions about getSelection().
	assert_not_equals(document.defaultView, null,
		"Sanity check: document.defaultView must not be null");

	assert_true(document.getSelection() instanceof Selection);
}, "document.getSelection() instanceof Selection");

test(function() {
	assert_not_equals(document.defaultView, null,
		"Sanity check: document.defaultView must not be null");
	assert_equals(document.getSelection(), document.getSelection());
}, "document.getSelection() === document.getSelection()");

test(function() {
	assert_not_equals(document.defaultView, null,
		"Sanity check: document.defaultView must not be null");
	assert_equals(window.getSelection(), document.getSelection());
}, "window.getSelection() === document.getSelection()");

// "Each selection is associated with a single range, which may be null and is
// initially null."
//
// "The rangeCount attribute must return 0 if the context object's range is
// null, otherwise 1."
test(function() {
	assert_equals(window.getSelection().rangeCount, 0,
		"window.getSelection().rangeCount must initially be 0");
	assert_equals(document.getSelection().rangeCount, 0,
		"document.getSelection().rangeCount must initially be 0");
}, "Selection's range must initially be null");

test(function() {
	var doc = document.implementation.createHTMLDocument("");
	assert_equals(doc.defaultView, null,
		"Sanity check: defaultView of created HTML document must be null");
	assert_equals(doc.getSelection(), null,
		"getSelection() on HTML document with null defaultView must be null");
}, "HTML document with no browsing context");

test(function() {
	var xmlDoc = document.implementation.createDocument(null, "", null);

	assert_true("getSelection" in xmlDoc, "XML document must have getSelection()");

	assert_equals(xmlDoc.defaultView, null,
		"Sanity check: defaultView of created XML document must be null");
	assert_equals(xmlDoc.getSelection(), null,
		"getSelection() on XML document with null defaultView must be null");
}, "XML document with no browsing context");

var iframe = document.createElement("iframe");
document.body.appendChild(iframe);
test(function() {
	assert_true("Selection" in iframe.contentWindow,
		"Sanity check: window must have Selection property");

	assert_true(iframe.contentWindow.getSelection() instanceof iframe.contentWindow.Selection,
		"window.getSelection() instanceof Selection");
	assert_not_equals(iframe.contentDocument.defaultView, null,
		"Sanity check: document.defaultView must not be null");
	assert_true(iframe.contentDocument.getSelection() instanceof iframe.contentWindow.Selection,
		"document.getSelection() instanceof Selection");
	assert_equals(iframe.contentWindow.getSelection(), iframe.contentDocument.getSelection(),
		"window.getSelection() === document.getSelection()");
	assert_not_equals(iframe.contentWindow.getSelection(), getSelection(),
		"getSelection() inside and outside iframe must return different objects");
}, "In an iframe");

test(function() {
	var doc = iframe.contentDocument.implementation.createHTMLDocument("");
	assert_equals(doc.defaultView, null,
		"Sanity check: defaultView of created HTML document must be null");
	assert_equals(doc.getSelection(), null,
		"getSelection() on HTML document with null defaultView must be null");
}, "HTML document with no browsing context inside an iframe");
document.body.removeChild(iframe);
</script>
